TiDB order-management-app示例订单ID数据类型修改为long类型

+++ demo-tidb/demo-order-management-app/common/src/main/java/com/future/demo/dto/OrderDTO.java
@@ -6,21 +6,12 @@ import com.future.demo.entity.DeleteStatus;
 import com.future.demo.entity.Status;
 import lombok.Data;
 
-import java.math.BigInteger;
 import java.time.LocalDateTime;
 import java.util.List;
 
 @Data
 public class OrderDTO {
-    // long 类型
-    /*private Long id;*/
-    // int 类型
-    /*private Integer id;*/
-    // biginteger 类型
-    private BigInteger id;
-    // uuid string 类型
-    /*private String id;*/
-
+    private Long id;
     private Long userId;
     @JsonFormat(pattern = "yyyy-MM-dd HH:mm:ss", timezone = "GMT+8")
     private LocalDateTime createTime;
