+++ demo-tidb/demo-order-management-app/common/src/main/java/com/future/demo/entity/IdCacheAssistantEntity.java
@@ -2,18 +2,8 @@ package com.future.demo.entity;
 
 import lombok.Data;
 
-import java.math.BigInteger;
-
 @Data
 public class IdCacheAssistantEntity {
     private Long id;
-
-    // long 类型
-    /*private Long orderId;*/
-    // int 类型
-    /*private Integer orderId;*/
-    // biginteger 类型
-    private BigInteger orderId;
-    // uuid string 类型
-    /*private String orderId;*/
+    private Long orderId;
 }
